refactor(pipeline/pii): split Identifier redactors (Account/Card/IsraeliId) into per-category modules
Phase 6 commit 2/7 of PiiRedactor split.

- Carve account/card/israeliId from src/Scrapers/Pipeline/Types/PiiRedactor.ts.
- New files: src/Scrapers/Pipeline/Types/PiiRedactor/{Account,Card,IsraeliId}.ts.
- Each module hosts CATEGORY descriptor + redact() + legacy
  redactAccount/redactCard/redactIsraeliId alias for back-compat.
- Helpers terminalSegment + lastSeparatorIndex move with Account
  (the only consumer).
- PiiRedactor.ts re-exports the legacy names from the new modules and
  imports them where dispatchStrategy/maskPathSegmentIfId need them.
- Tests: zero body diff — existing 138 PiiRedactor tests pass
  unchanged (re-exports preserve byte-identical behaviour).

Verified:
- npm run build green
- npm run test:pipeline (PiiRedactor cluster) green
- npm run lint + npm run type-check green
- Each new file ≤ 80 LoC

Master plan: pipeline-decoupling-master-2026-05-28 / phase-6
